adds new requested changes
