+++ src/functions/memo-management/drive-memos/get-content-pdf/index.ts
@@ -1,0 +1,16 @@
+import { handlerPath } from "@libs/handler-resolver";
+
+export default {
+  handler: `${handlerPath(__dirname)}/handler.main`,
+  events: [
+    {
+      httpApi: {
+        method: "get",
+        path: "/google-drive/memos-content",
+        authorizer: {
+          name: "timebankKeycloakAuthorizer"
+        }
+      }
+    }
+  ],
+}
