+++ src/functions/memo-management/trello-cards/delete-trello-card/handler.ts
@@ -1,5 +1,5 @@
 import { APIGatewayProxyEvent, APIGatewayProxyHandler } from "aws-lambda";
-import { TrelloService } from "src/database/services/trello-api-service";
+import { TrelloService } from "src/service/trello-api-service";
 import { middyfy } from "src/libs/lambda";
 
 /**
