+++ src/functions/memo-management/trello-cards/get-board-members/handler.ts
@@ -1,5 +1,5 @@
 import { APIGatewayProxyHandler } from "aws-lambda";
-import { TrelloService } from "src/database/services/trello-api-service";
+import { TrelloService } from "src/service/trello-api-service";
 import { middyfy } from "src/libs/lambda";
 
 /**
@@ -10,11 +10,11 @@ import { middyfy } from "src/libs/lambda";
 const getBoardMembersHandler: APIGatewayProxyHandler = async () => {
   try {
     const trello = new TrelloService();
-    const emails = await trello.getBoardMembers();
+    const members = await trello.getBoardMembers();
 
     return {
       statusCode: 200,
-      body: JSON.stringify(emails),
+      body: JSON.stringify(members),
     };
   } catch (error) {
     return {
